pyproject.toml
[project]
name = "project-ecommerce-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
dependencies = [
    "chromadb>=1.3.3",
    "groq>=0.33.0",
    "pandas>=2.3.3",
    "pandasql>=0.7.3",
    "python-dotenv>=1.2.1",
    "selenium>=4.38.0",
    "semantic-router>=0.1.11",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
]
